requirements.txt
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.4.0
matplotlib>=3.8.0
seaborn>=0.13.0
jupyter>=1.0.0
pyyaml>=6.0
requests>=2.30.0
tqdm>=4.65.0
fastapi>=0.100.0
uvicorn>=0.22.0